Extracted the virtual response
