+++ AngleSharp/AngleSharp.Core.csproj
@@ -256,6 +256,7 @@
     <Compile Include="Network\RequestProcessors\ImageRequestProcessor.cs" />
     <Compile Include="Network\RequestProcessors\IRequestProcessor.cs" />
     <Compile Include="Network\RequestProcessors\StyleSheetRequestProcessor.cs" />
+    <Compile Include="Network\VirtualResponse.cs" />
     <Compile Include="Parser\Css\CssBuilder.cs" />
     <Compile Include="Parser\Css\CssParserOptions.cs" />
     <Compile Include="Parser\Css\Tokens\CssColorToken.cs" />
